fix for import in tests
