+++ client/src/app/[tenant]/agents/[taskId]/[taskSchemaId]/proxy-playground/proxy-messages/ProxyMessagesView.test.ts
@@ -1,4 +1,4 @@
-import { elementIdForMessage } from './ProxyMessagesView';
+import { elementIdForMessage } from './elementIdForMessage';
 
 describe('elementIdForMessage', () => {
   const userMsg = (overrides = {}) => ({ role: 'user' as const, content: [], ...overrides });
